requirements.txt
Django
Djangorestframework
Djangorestframework-simplejwt
numpy
psycopg2-binary